# ARROW-3548: [Plasma] Add CreateAndSeal object store method for faster puts for small objects.