--- a/cpp/src/plasma/protocol.h
+++ b/cpp/src/plasma/protocol.h
@@ -20,6 +20,7 @@
 
 #include <memory>
 #include <unordered_map>
+#include <string>
 #include <vector>
 
 #include "arrow/status.h"