--- a/python/pyarrow/tests/test_plasma.py
+++ b/python/pyarrow/tests/test_plasma.py
@@ -217,19 +217,19 @@ class TestPlasmaClient(object):
         for i in range(1000):
             object_id = random_object_id()
             object_ids.append(object_id)
-            self.plasma_client.create_and_seal(object_id, i * b'a', i * b'a')
+            self.plasma_client.create_and_seal(object_id, i * b'a', i * b'b')
 
         for i in range(1000):
             assert self.plasma_client.get_buffer(object_ids[i]) == i * b'a'
             assert (self.plasma_client.get_metadata(
                         [object_ids[i]])[0].to_pybytes()
-                    == i * b'a')
+                    == i * b'b')
 
         # Make sure that creating the same object twice raises an exception.
         object_id = random_object_id()
-        self.plasma_client.create_and_seal(object_id, b'a', b'a')
+        self.plasma_client.create_and_seal(object_id, b'a', b'b')
         with pytest.raises(pa.PlasmaObjectExists):
-            self.plasma_client.create_and_seal(object_id, b'a', b'a')
+            self.plasma_client.create_and_seal(object_id, b'a', b'b')
 
         # Make sure that these objects can be evicted.
         big_object = DEFAULT_PLASMA_STORE_MEMORY // 10 * b'a'