To benchmark, start a store with `plasma_store_server -s /tmp/store -m 1000000000`

Then do

```python
import pyarrow.plasma as plasma
client = plasma.connect('/tmp/store', '', 0)
```

```python
def before():
    object_id = plasma.ObjectID.from_random()
    client.create(object_id, 20, b'')
    client.seal(object_id)

def after():
    object_id = plasma.ObjectID.from_random()
    client.create_and_seal(object_id, 20 * b'a', b'')
```

```
%timeit before()
63.4 µs ± 631 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit after()
30.5 µs ± 669 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

There's actually more that could be done in the future in order to not have to wait for the return IPC if the client "reserves" a big chunk of memory up front.